Refactored several couples of coordinates into Vec2i occurrences
Added some Vec2i operators to ease their manipulation
Removed some zombie code

+++ src/include/viewport.h
@@ -96,7 +96,7 @@ public:
 
 	bool Contains(const PixelPos &screenPos) const;
 
-	void Restrict(int &screenPosX, int &screenPosY) const;
+	void Restrict(short &screenPosX, short &screenPosY) const;
 
 	PixelSize GetPixelSize() const;
 	const PixelPos &GetTopLeftPos() const { return TopLeftPos;}
